### 2017-07-23 #329 @bjmorgan
Potential reviewers:

From the JOSS list:
- khinsen
- jochym

Possible other reviewers:
- Adam Jackson: https://github.com/ajjackson
- [Aron Walsh](http://www.imperial.ac.uk/people/a.walsh): https://github.com/aronwalsh
- [Paul Erhart](https://www.chalmers.se/en/staff/Pages/Paul-Erhart.aspx): https://gitlab.com/erhart
