### 2017-07-23 #329 @labarba
Thank you , @bjmorgan, for suggesting reviewers… most helpful!

I was about to send an email to request a review from Dr. Kristin A. Persson (or someone she might recommend). Seem appropriate?
http://perssongroup.lbl.gov/people.html
