### 2017-07-23 #329 @labarba
In the meantime, I wonder if you could take the time to expand a little the description in the paper. We (the JOSS editors) are in the process of adding more guidance for authors, including a better indication of expected paper length. We think it's hard to get a picture of the functionality of the software, and also include a statement of need, in less than 250 words. One reason is that JOSS is welcoming software from a wide variety of fields, and we need the papers to illustrate the purpose of the software, for diverse, non-specialist readers.
