### 2017-07-24 2abfc47
Faster SymmetryOperation.operate_on method
