+++ bsym/symmetry_operation.py
@@ -120,8 +120,7 @@ class SymmetryOperation:
             the new configuration obtained by operating on configuration with this symmetry operation. 
         """
         # returns the site occupation vector generated by applying this symmetry operation
-        new_configuration = Configuration( ( self.matrix * configuration ).tolist() )
-        return( new_configuration )
+        return Configuration( np.array( self.matrix * configuration ) )
 
     def character( self ):
         """
