### 2017-07-24 847fe97
unique_permutations now evaluates correctly inside list()
