### 2017-07-24 1aba77e
Configuration class no longer subclasses matrix
