+++ tests/test_configuration.py
@@ -4,55 +4,52 @@ from bsym.configuration import Configuration
 from bsym import SymmetryOperation
 import numpy as np
 
-def column_vector( l ):
-    return np.array( [ l ] ).T
-
 class TestConfiguration( unittest.TestCase ):
 
     def setUp( self ):
-        self.configuration = Configuration( column_vector( [ 1, 0, 0 ] ) )
+        self.configuration = Configuration( [ 1, 0, 0 ] )
 
     def test_matches_returns_true_for_a_match( self ):
         self.assertEqual( self.configuration.matches( self.configuration ), True )
 
     def test_matches_returns_false_for_a_non_match( self ):
-        other_configuration = Configuration( column_vector( [0, 0, 1] ) )
+        other_configuration = Configuration( [0, 0, 1] )
         self.assertEqual( self.configuration.matches( other_configuration ), False )
 
     def test_is_equivalent_to_if_equivalent( self ):
-        test_configuration = Configuration( column_vector( [0, 1, 0] ) )
+        test_configuration = Configuration( [0, 1, 0] )
         symmetry_operations = [ Mock( spec=SymmetryOperation ) ]
         symmetry_operations[0].operate_on = Mock( return_value=test_configuration )
         self.assertEqual( self.configuration.is_equivalent_to( test_configuration, symmetry_operations ), True )
 
     def test_is_equivalent_to_if_not_equivalent( self ):
-        test_configuration = Configuration( column_vector( [ 0, 1, 0 ] ) )
+        test_configuration = Configuration( [ 0, 1, 0 ] )
         symmetry_operations = [ Mock( spec=SymmetryOperation ) ]
-        symmetry_operations[0].operate_on = Mock( return_value=Configuration( column_vector( [ 0, 0, 1 ] ) ) )
+        symmetry_operations[0].operate_on = Mock( return_value=Configuration( [ 0, 0, 1 ] ) )
         self.assertEqual( self.configuration.is_equivalent_to( test_configuration, symmetry_operations ), False )
 
     def test_is_in_list( self ):
-        configuration_list = [ Configuration( column_vector( [ 0, 1, 0 ] ) ), 
-                               Configuration( column_vector( [ 1, 0, 0 ] ) ) ]
+        configuration_list = [ Configuration( [ 0, 1, 0 ] ), 
+                               Configuration( [ 1, 0, 0 ] ) ]
         self.configuration.matches = Mock( return_value=True )
         self.assertEqual( self.configuration.is_in_list( configuration_list ), True )
 
     def test_is_in_list_fails( self ):
-        configuration_list = [ Configuration( column_vector( [ 0, 1, 0 ] ) ), 
-                               Configuration( column_vector( [ 1, 0, 0 ] ) ) ]
+        configuration_list = [ Configuration( [ 0, 1, 0 ] ), 
+                               Configuration( [ 1, 0, 0 ] ) ]
         self.configuration.matches = Mock( return_value=False )
         self.assertEqual( self.configuration.is_in_list( configuration_list ), False )
 
     def test_has_equivalent_in_list( self ):
-        configuration_list = [ Configuration( column_vector( [ 0, 1, 0 ] ) ), 
-                               Configuration( column_vector( [ 1, 0, 0 ] ) ) ]
+        configuration_list = [ Configuration( [ 0, 1, 0 ] ), 
+                               Configuration( [ 1, 0, 0 ] ) ]
         symmetry_operations = [ Mock( spec=SymmetryOperation ) ]
         self.configuration.is_equivalent_to = Mock( return_value=True )
[…]
