### 2017-07-24 069cf83
SpaceGroup uses np.ndarray instead of np.matrix
