### 2017-07-24 b3f3b00
Added type checking for Configuration.matches
