### 2017-07-24 13b99cd
removed numba imports
