### 2017-07-24 7906949
removed @jit decorators
