### 2017-07-24 2e5175f
Better variable names for lists
