### 2017-07-25 e6c08ed
Explain count attribute in notebook
