### 2017-07-25 a8aa408
Attempted fix for symmetry analysis in non-cubic cells
