### 2017-07-25 97d00d6
Added verbose flags to some pymatgen interface functions
