### 2017-07-25 25eca09
Expanded description in paper.md
