+++ paper.bib
@@ -31,5 +31,106 @@
 }
 
 
+@article{LudwigEtAl_PhysRevB2011,
+	Author = {Ludwig, Christian D. R. and Gruhn, Thomas and Felser, Claudia and Windeln, Johannes},
+	Date-Added = {2017-07-25 12:49:40 +0000},
+	Date-Modified = {2017-07-25 12:49:54 +0000},
+	Journal = {Phys. Rev. B},
+	Month = may,
+	Number = {17},
+	Pages = {622--8},
+	Title = {Defect structures in {C}u{I}n{S}e$_2$: A combination of Monte Carlo simulations and density functional theory},
+	Volume = {83},
+	Year = {2011}}
 
+
+@article{GanoseAndScanlon_JMaterChemA2016,
+	Author = {Ganose, Alex M. and Scanlon, David O.},
+	Date-Added = {2017-07-25 10:59:19 +0000},
+	Date-Modified = {2017-07-25 10:59:32 +0000},
+	Journal = {J. Mater. Chem. A},
+	Month = feb,
+	Pages = {1467--1475},
+	Title = {Band gap and work function tailoring of {S}n{O}$_2$ for improved transparent conducting ability in photovoltaics},
+	Volume = {4},
+	Year = {2016}}
+
+@article{MorganAndWatson_JPhysChemLett2011,
+	Author = {Morgan, Benjamin J. and Watson, Graeme W.},
+	Date-Added = {2012-04-02 17:49:13 +0000},
+	Date-Modified = {2015-11-16 08:26:50 +0000},
+	Journal = {J. Phys. Chem. Lett.},
+	Month = jun,
+	Pages = {1657---1611},
+	Title = {The Role of Lithium Ordering in the {L}i$_x${T}i{O}$_2$ Anatase $\to$ Titanate Phase Transition},
+	Volume = {2},
+	Year = {2011}}
+
+@article{GrieshammerEtAl_PhysChemChemPhys2014,
+	Author = {Grieshammer, Steffen and Grope, Benjamin O. H. and Koettgen, Julius and Martin, Manfred},
+	Date-Added = {2017-07-25 13:19:16 +0000},
+	Date-Modified = {2017-07-25 13:19:32 +0000},
+	Journal = {Phys. Chem. Chem. Phys.},
+	Number = {21},
+	Pages = {9974--13},
+	Title = {A combined {DFT}+$U$ and {M}onte {C}arlo study on rare earth doped ceria},
+	Volume = {16},
+	Year = {2014}}
+
+@article{DaltonEtAl_ChemMater2012,
+	Author = {Dalton, Andrew S. and Belak, Anna A. and Van Der Ven, Anton},
+	Date-Added = {2012-05-22 16:18:00 +0000},
+	Date-Modified = {2012-05-22 16:18:24 +0000},
+	Journal = {Chem. Mater.},
+	Month = apr,
+	Pages = {1568--1574},
+	Title = {Thermodynamics of Lithium in {T}i{O}$_2$({B}) from First Principles},
+	Volume = {24},
+	Year = {2012}}
+
+
+@article{ZungerEtAl_PhysRevLett1990,
+	Author = {Zunger, A. and Wei, S.H. and Ferreria, L. G. and Bernard, J. E.},
+	Date-Added = {2017-07-25 06:16:03 +0000},
+	Date-Modified = {2017-07-25 06:16:38 +0000},
+	Journal = {Phys. Rev. Lett.},
+	Number = {3},
+	Pages = {353--356},
+	Title = {Special Quasirandom Structures},
+	Volume = {65},
+	Year = {1990}}
+
[…]
