### 2017-07-25 e583c5d
Typo fixed in paper.md
