### 2017-07-25 939e43c
Added bsym reference to paper.md
