### 2017-07-25 #329 @bjmorgan
@labarba I have rewritten my `paper.md` to give a more detailed explanation of the intended purpose of the software.

https://github.com/bjmorgan/bsym/commit/25eca0997802ae5cbe72ea29875c5204dcbdb6e2
