### 2017-07-28 de6899e
Fixed typo in README
