### 2017-08-02 af2a38b
setup.py uses pandoc to parse README
