### 2017-08-02 67c5991
setup.py now actually sets up
