### 2017-08-02 4c228b8
Added PyPI & pip info
