### 2017-08-03 78d94f8
Fixed typo in README.md
