### 2017-08-17 #370 @whedon
[REVIEW]: bsym - a basic symmetry module
**Submitting author:** @bjmorgan (<a href="http://orcid.org/0000-0002-3056-8233">Benjamin J. Morgan</a>)
**Repository:** <a href="https://github.com/bjmorgan/bsym" target ="_blank">https://github.com/bjmorgan/bsym</a>
**Version:** v1.0.b1

- [x] **Version:** Does the release version given match the GitHub release (v1.0.b1)?
