### 2017-08-17 bjmorgan/bsym#1 @aronwalsh
Missing module
`bsym.interface.pymatgen` is used in the example notebook, but does not appear to be in the current pip release (it is in the GitHub version)
