### 2017-08-17 bjmorgan/bsym#2 @aronwalsh
Configuration degeneracy 
In the materials notebook example, it is not an error, but the configuration degeneracy depends on the route:

Sequential substitutions:
[4, 4, 4, 1, 2]

Simultaneous substitutions:
[64, 64, 64, 16, 32]

The latter should be more relevant for statistical analysis / thermodynamics for materials, which should be mentioned.
