### 2017-08-17 bjmorgan/bsym#3 @aronwalsh
Group theory 
As I understand there is no internal library of point group operations or character tables included. Could these be imported from spglib?
