### 2017-08-17 #370 @aronwalsh
bsym is a python based symmetry module with applications in the molecular and materials domain. The code installed simply using `pip3 install bsym` and tests completed using `python3 -m unittest discover`. I found the source code to well written and documented. I particularly liked the jupyter notebook example, which introduces the background and functionality.  The code is useful pedagogically and could be employed for teaching group theory, but also for practical research in materials science.  I raised three minor issues on the repository tracker: https://github.com/bjmorgan/bsym/issues
