### 2017-08-17 #370 @labarba
Thanks for the lightning-fast review, @aronwalsh !!
I see that you have checked off all the items in the reviewer checklist. Do I take it that your recommendation is to _Accept_ the submission, at this point? That is, we trust the authors to address the issues you posted on their repo, as they are minor, and proceed?
