### 2017-08-17 #370 @labarba
@arfon : This submission is ready for acceptance.

@aronwalsh : Thank you for your contribution to JOSS, and do consider submitting your research software!

@bjmorgan : Your next step is to make the software archive (if you haven't already done so), and please post the DOI of the archive here.
