### 2017-08-17 bjmorgan/bsym#1 @bjmorgan
`bsym.interface` was missing from `setup.py`. This should now install the `interface` submodule correctly when installing from PyPI via `pip`.

relevant commit: ccc463e8318c6dfc45e90fad19d20e25ea5e4ffa

new release: https://github.com/bjmorgan/bsym/releases/tag/1.0.b6
