### 2017-08-17 f8048f7
Added full_configuration_degeneracy attribute
pymatgen Structure attributes generated via
the unique_structure_substitutions function
now have a full_configuration_degeneracy
attribute, which included degeneracy terms
from any previous symmetry-breaking substitutions.

Changes made in response to
https://github.com/bjmorgan/bsym/issues/2
