+++ tests/test_interface_pymatgen.py
@@ -90,6 +90,37 @@ class TestPymatgenInterface( unittest.TestCase ):
         distances = np.array( sorted( [ s.get_distance( s.indices_from_symbol('O')[0], s.indices_from_symbol('Na')[0] ) for s in ns ] ) )
         np.testing.assert_array_almost_equal( distances, np.array( [ 0.75    ,  1.06066 ,  1.5     ,  1.677051,  2.12132 ] ) )
         np.testing.assert_array_equal( np.array( sorted( [ s.number_of_equivalent_configurations for s in ns ] ) ), np.array( [ 1, 2, 4, 4, 4 ] ) )
+        np.testing.assert_array_equal( np.array( sorted( [ s.full_configuration_degeneracy for s in ns ] ) ), np.array( [ 1, 2, 4, 4, 4 ] ) )
+
+    def test_unique_structure_substitutions_in_two_steps_gives_full_degeneracies( self ):
+        # integration test
+        # Create a pymatgen structure with 16 sites in a 4x4 square grid
+        coords = np.array( [ [ 0.0, 0.0, 0.0 ],
+                             [ 0.25, 0.0, 0.0 ],
+                             [ 0.5, 0., 0.0 ],
+                             [ 0.75, 0.0, 0.0 ],
+                             [ 0.0, 0.25, 0.0 ],
+                             [ 0.25, 0.25, 0.0 ],
+                             [ 0.5, 0.25, 0.0 ],
+                             [ 0.75, 0.25, 0.0 ],
+                             [ 0.0, 0.5, 0.0 ],
+                             [ 0.25, 0.5, 0.0 ],
+                             [ 0.5, 0.5, 0.0 ],
+                             [ 0.75, 0.5, 0.0 ],
+                             [ 0.0, 0.75, 0.0 ],
+                             [ 0.25, 0.75, 0.0 ],
+                             [ 0.5, 0.75, 0.0 ],
+                             [ 0.75, 0.75, 0.0 ] ] )
+        atom_list = [ 'Li' ] * len( coords )
+        lattice = Lattice.from_parameters( a = 3.0, b=3.0, c=3.0, alpha=90, beta=90, gamma=90 )
+        parent_structure = Structure( lattice, atom_list, coords )
+        us = unique_structure_substitutions( parent_structure, 'Li', { 'Na':1, 'Li':15 } )
+        ns = unique_structure_substitutions( us[0], 'Li', { 'Mg':1, 'Li':14 } )
+        self.assertEqual( len( ns ), 5 )
+        distances = np.array( sorted( [ s.get_distance( s.indices_from_symbol('Mg')[0], s.indices_from_symbol('Na')[0] ) for s in ns ] ) )
+        np.testing.assert_array_almost_equal( distances, np.array( [ 0.75    ,  1.06066 ,  1.5     ,  1.677051,  2.12132 ] ) )
+        np.testing.assert_array_equal( np.array( sorted( [ s.number_of_equivalent_configurations for s in ns ] ) ), np.array( [ 1, 2, 4, 4, 4 ] ) )
+        np.testing.assert_array_equal( np.array( sorted( [ s.full_configuration_degeneracy for s in ns ] ) ), np.array( [ 16, 32, 64, 64, 64 ] ) )
 
     def test_unique_structure_substitutions_with_mismatched_site_distribution_raises_ValueError( self ):
         mock_structure = Mock( spec=Structure )
@@ -100,7 +131,9 @@ class TestPymatgenInterface( unittest.TestCase ):
     def test_pymatgen_structure_can_be_patched( self ):
         with self.assertRaises( AttributeError ):
             self.structure.number_of_equivalent_configurations
-    
+        with self.assertRaises( AttributeError ):
[…]
