### 2017-08-17 bjmorgan/bsym#2 @bjmorgan
I have updated the `unique_structure_substitutions` function in `bsym.interface.pymatgen`. `Structure` objects generated now have two additional attributes:
`number_of_equivalent_configurations` which gives the degeneracy under that substitution, and
`full_configuration_degeneracy` which records the cumulative product of degeneracies from sequential substitutions.

I have updated the example in the Jupyter notebook to explain the difference between these two values.

relevant commit: f8048f7d2b372e2540e87e74d363f7a4e88e705d
