### 2017-08-17 bjmorgan/bsym#3 @bjmorgan
@aronwalsh It isn't clear why this would be useful. Can you provide an example use case?
