### 2017-08-17 #370 @bjmorgan
@aronwalsh Thank you for the speedy review.
@labarba I will address the issues raised in the review, and then generate an archived release.
