### 2017-08-17 bjmorgan/bsym#3 @aronwalsh
I had the lazy user in mind, for faster application to chemical problems. The ability to call a set of predefined symmetry operations instead of having to define them.
