### 2017-08-17 bjmorgan/bsym#3 @bjmorgan
If the system is described by a `pymatgen` `Structure` object, you can generate the `bsym` `SpaceGroup` using `bsym.interface.spacegroup_from_structure()`.

http://bsym.readthedocs.io/en/latest/api/interface/pymatgen.html#bsym.interface.pymatgen.spacegroup_from_structure

```python
def spacegroup_from_structure( structure, subset=None ):
    """
    Generates a ``SpaceGroup`` object from a `pymatgen` ``Structure``.

    Args:
        structure (pymatgen ``Structure``): structure to be used to define the :any:`SpaceGroup`.
        subset    (Optional [list]):        list of atom indices to be used for generating the symmetry operations.

    Returns:
        a new :any:`SpaceGroup` instance
    """
```

The design choice to use the `pymatgen` API, rather than reimplement a direct interface with `spglib`, was, in part, motivated by the excellent support in `pymatgen` for converting between multiple standard structure file and data formats, and the expectation that most users would prefer to work with structures in a "standard" format, rather than having to generate a `bsym` structure object.
