### 2017-08-18 db69b79
Added configuration_space_from_structure()
