### 2017-08-18 9636aa9
Added space_group_symbol_from_structure()
