### 2017-08-18 58d0a42
Numeric list for ConfigurationSpace objects
