### 2017-08-18 d05267b
Added more pymatgen interface examples
