### 2017-08-18 8dbe909
Added tests for raising TypeError
