### 2017-08-18 af3c700
Added error checking unittests
