### 2017-08-18 dd236aa
Added unittest for type checking
