### 2017-08-18 d981e68
Added unittest for CoordinateConfigSpace
