### 2017-08-18 19ef10a
Reverted to ImportError for 3.5 compatibility
