### 2017-08-18 bjmorgan/bsym#3 @bjmorgan
For better discoverability, I have added examples of the `spacegroup_from_structure()` and `configuration_space_from_structure()` functions from `bsym.interface.pymatgen` to the example Jupyter notebook.

relevant commit: d05267b3aaee857bd99e1febff7268a348610e4e
