### 2017-08-18 #370 @bjmorgan
The issues raised by @aronwalsh have been responded to / addressed:

https://github.com/bjmorgan/bsym/issues/1
https://github.com/bjmorgan/bsym/issues/2
https://github.com/bjmorgan/bsym/issues/3

Changes:
- A bug, where submodules were not installing correctly using `pip`, has been fixed.
- The full configuration degeneracy is now preserved over a series of site substitutions.
- Examples using the `pymatgen` interface to auto-generate `SymmetryGroup` and `ConfigurationSpace` objects have been added to the example Jupyter notebook.
