### 2017-08-18 f6c2867
Bumped version to 1.0
