### 2017-08-18 338e549
Bumped version to 1.0.0
