### 2017-08-18 e86ddea
Added JOSS status badge to README
