### 2017-08-18 #370 @bjmorgan
@labarba

- DOI of the submitted version, v1.0.b1: [10.5281/zenodo.833875](https://doi.org/10.5281/zenodo.833875)
- DOI of the "as accepted" version, v1.0.1: [10.5281/zenodo.845328](https://doi.org/10.5281/zenodo.845328)
