### 2017-08-18 #370 @arfon
@bjmorgan - it looks like there are a couple of issues with your paper with the citations, can you check the bibtex entries are correctly named please:

```
pandoc-citeproc: reference MorganEtAl_JMaterChemA2016 not found
pandoc-citeproc: reference LurchEtAl_ModelSimulMaterSciEng2009 not found
```

Also, there's a typo in the metadata of your paper. The first line is quoted with a beginning `'` and ending `"` which needs fixing:

```
title: 'bsym: A basic symmetry module"
```
