### 2017-08-18 9f326e2
Typos and citations fixed for paper.md
in response to
https://github.com/openjournals/joss-reviews/issues/370#issuecomment-323314045
