+++ paper.bib
@@ -134,3 +134,29 @@
 	Volume = {25},
 	Year = {2013}}
 
+
+@article{MorganEtAl_JMaterChemA2016,
+	Author = {Morgan, Benjamin J. and Carrasco, Javier and Teobaldi, Gilberto},
+	Date-Added = {2017-04-03 11:28:26 +0000},
+	Date-Modified = {2017-04-03 11:28:26 +0000},
+	Journal = {J. Mater. Chem. A},
+	Month = oct,
+	Number = {43},
+	Pages = {17180--17192},
+	Title = {Variation in surface energy and reduction drive of a metal oxide lithium-ion anode with stoichiometry: a {DFT} study of lithium titanate spinel surfaces},
+	Volume = {4},
+	Year = {2016}}
+
+
+@article{LerchEtAl_ModellingSimulMaterSciEng2009,
+	Author = {Lerch, D. and Wieckhorst, O. and Hart, G. L. W. and Forcade, R. W. and M{\"u}ller, S.},
+	Date-Added = {2017-08-18 10:07:39 +0000},
+	Date-Modified = {2017-08-18 10:07:39 +0000},
+	Journal = {Modelling Simul. Mater. Sci. Eng.},
+	Month = jun,
+	Number = {5},
+	Pages = {055003--20},
+	Title = {{UNCLE}: a code for constructing cluster expansions for arbitrary lattices with minimal user-input},
+	Volume = {17},
+	Year = {2009}}
+
