### 2017-08-18 #370 @bjmorgan
@arfon These should be fixed now. `pandoc --filter pandoc-citeproc paper.md -o paper.pdf` runs without errors.
